[MOV] website, html_builder, *: move many2one_option to html_builder
*: website_blog

task-4367641
