+++ addons/html_builder/static/src/plugins/many2one_option.js
@@ -3,7 +3,7 @@ import { onWillStart } from "@odoo/owl";
 import { useService } from "@web/core/utils/hooks";
 
 export class Many2OneOption extends BaseOptionComponent {
-    static template = "website_builder.Many2OneOption";
+    static template = "html_builder.Many2OneOption";
     static props = [];
     setup() {
         super.setup();
