+++ addons/html_builder/static/src/plugins/many2one_option.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <templates xml:space="preserve">
 
-<t t-name="website_builder.Many2OneOption">
+<t t-name="html_builder.Many2OneOption">
     <BuilderRow label="label">
         <BuilderMany2One action="'many2One'" model="model" fields="['name']" allowUnselect="false"/>
     </BuilderRow>
