+++ addons/html_builder/static/src/plugins/many2one_option_plugin.js
