+++ addons/website_blog/static/src/website_builder/author_avatar_many2one_plugin.js
@@ -1,5 +1,5 @@
 import { patch } from "@web/core/utils/patch";
-import { Many2OneAction } from "@website/builder/plugins/options/many2one_option_plugin";
+import { Many2OneAction } from "@html_builder/plugins/many2one_option_plugin";
 
 patch(Many2OneAction, {
     dependencies: [...Many2OneAction.dependencies, "history"],
